requirements.txt
streamlit==1.52.0
snowflake-ml-python==1.20.0
snowflake-snowpark-python==1.44.0
watchdog
